Bug fix in classifiers.py: cannot convert "?" to float
Missing values are marked in the dataset as question marks(?). When reading the csv file, we could not convert the question mark into a float like expected. Since question mark represents a missing data, we will use a policy of removing all examples/instances which contain missing data. This commit contains the implementation of this new method for removing missing-data instances